+++ controllers/categoriesController.js
@@ -35,17 +35,7 @@ const categoriesController = {
     const items = await getItems(req.params);
     res.render("category", { title: category, category, categories, items });
   }),
-  getCategoryEdit: asyncHandler(async (req, res) => {
-    console.log("getCategoryEdit running...");
-    console.log("req.url:", req.url);
-    console.log("req.params:", req.params);
-    const categories = await getCategories();
-    const { category } = req.params;
-    const input = { category };
-    const edit = true;
-    // Need to include category form
-    res.render("edit", { title: "Edit Category", categories, edit, input });
-  }),
+  getCategoryEdit: asyncHandler(async (req, res) => {}),
   getCategoryDelete: asyncHandler(async (req, res) => {
     console.log("getCategoryDelete running...");
     console.log("req.url:", req.url);
@@ -53,13 +43,6 @@ const categoriesController = {
     const categories = await getCategories();
     const { category } = req.params;
     res.render("delete", { title: "Delete Category", categories });
-  }),
-  postCategoryEdit: asyncHandler(async (req, res) => {
-    console.log("postCategoryEdit running...");
-    console.log("req.url:", req.url);
-    console.log("req.query:", req.query);
-    console.log("req.params:", req.params);
-    console.log("req.body:", req.body);
   }),
   postCategoryDelete: asyncHandler(async (req, res) => {
     console.log("postCategoryDelete running...");