+++ routes/categoriesRouter.js
@@ -2,14 +2,10 @@ const { Router } = require("express");
 const {
   getCategoryItems,
   getCategories,
-  getCategoryEdit,
-  getCategoryDelete,
 } = require("../controllers/categoriesController");
 
 const categoriesRouter = new Router();
 
-categoriesRouter.get("/edit/:category", getCategoryEdit);
-categoriesRouter.get("/delete/:category", getCategoryDelete);
 categoriesRouter.get("/:category", getCategoryItems);
 categoriesRouter.get("/", getCategories);